requirements.txt
# Core
python-telegram-bot==21.6
python-dotenv==1.0.1

# Embeddings
sentence-transformers==3.1.1

# LLM backends (install only what you need)
openai==1.51.0           # needed only for LLM_BACKEND=openai
# transformers==4.44.2   # needed only for LLM_BACKEND=huggingface
# torch==2.4.0           # required by transformers backend

# Utilities
numpy==1.26.4
